compressed skill build table
