+++ public/js/yasp.js
@@ -91,7 +91,7 @@ window.tooltips = require('./tooltips.js');
 window.formatHtml = require("./formatHtml.js");
 window.createHistogram = require('./histograms.js');
 window.createCalHeatmap = require('./calheatmap.js');
-window.buildMap = require('./map.js');
+window.buildMap = require('./buildMap.js');
 window.playerMatches = require('./playerMatches.js');
 window.drawHeroes = require('./drawHeroes.js');
 window.drawTeammates = require('./drawTeammates.js');
